Move the deep link parser to its own file
