+++ src/actions/DeepLinkingActions.js
@@ -7,7 +7,7 @@ import { showError, showToast } from '../components/services/AirshipInstance.js'
 import { guiPlugins } from '../constants/plugins/GuiPlugins.js'
 import { EDGE_LOGIN, EXCHANGE_SCENE, PLUGIN_VIEW, WALLET_LIST_SCENE } from '../constants/SceneKeys.js'
 import s from '../locales/strings.js'
-import { type DeepLink } from '../types/DeepLink.js'
+import { type DeepLink } from '../types/DeepLinkTypes.js'
 import { type Dispatch, type GetState, type RootState } from '../types/reduxTypes.js'
 import { Actions } from '../types/routerTypes.js'
 import { activatePromotion } from './AccountReferralActions.js'
