+++ src/components/scenes/LoginScene.js
@@ -15,7 +15,7 @@ import edgeLogo from '../../assets/images/edgeLogo/Edge_logo_L.png'
 import { type ThemeProps, withTheme } from '../../components/services/ThemeContext.js'
 import s from '../../locales/strings.js'
 import { THEME } from '../../theme/variables/airbitz.js'
-import { type DeepLink } from '../../types/DeepLink.js'
+import { type DeepLink } from '../../types/DeepLinkTypes.js'
 import { connect } from '../../types/reactRedux.js'
 import { type GuiTouchIdInfo } from '../../types/types.js'
 import { showHelpModal } from '../modals/HelpModal.js'
